Added browser key checking
Session's now store the username of the current user. When a page is loaded, the session's user is checked with the currently logged in user, and if these do not match, the user is logged out. This is to ensure that the correct browser encryption key is loaded when a website is accessed.